rename gas_analysis and gas_plot to bulk_gas_analysis and bulk_gas_plotting rename AnalysisConfig to TweezerAnalysisConfig and add MOTAnalysisConfig. Some other edits to acommedate for the config structure.
