+++ singleshot/tweezer_survival_alternative_bkg.py
@@ -1,10 +1,12 @@
-from analysis_lib import TweezerAnalysis, TweezerPlotter, AnalysisConfig
+from .tweezer_analysis import TweezerAnalysis, TweezerPlotter
 import matplotlib.pyplot as plt
+from .analysis_config import kinetix_system, TweezerAnalysisConfig
+from .plot_config import PlotConfig
 
 # ROI config path: use when load_roi = False
 roi_config_path = 'X:\\userlib\\analysislib\\scripts\\multishot\\tweezer_roi.yaml'
 
-analysis_config = AnalysisConfig.from_yaml(roi_config_path)
+analysis_config = TweezerAnalysisConfig.from_yaml(roi_config_path)
 analysis_config.method = 'alternative'
 
 # Initialize analysis with background ROI and standard ROI loading
@@ -15,7 +17,8 @@ tweezer_analyzer = TweezerAnalysis(
 )
 
 tweezer_plotter = TweezerPlotter(
-    tweezer_analyzer
+    tweezer_analyzer,
+    plot_config=PlotConfig()
 )
 
 # plot results
